openGL

+++ Engine/Math.cpp
@@ -40,9 +40,9 @@ mat4f Math::buildRotateMatrixX(float rad)
 	mat4f m;
 	m.toIdentity();
 
-	m[1][1] = m[2][2]=  cos(rad);
-	m[1][2] = -sin(rad);
-	m[2][1] = sin(rad);
+	m[1][1] = m[2][2] = static_cast<float>(cos(rad));
+	m[1][2] = static_cast<float>(-sin(rad));
+	m[2][1] = static_cast<float>(sin(rad));
 
 	return m;
 }
@@ -52,9 +52,9 @@ mat4f Math::buildRotateMatrixY(float rad)
 	mat4f m;
 	m.toIdentity();
 
-	m[0][0] = m[2][2] = cos(rad);
-	m[2][0] = -sin(rad);
-	m[0][2] = sin(rad);
+	m[0][0] = m[2][2] = static_cast<float>(cos(rad));
+	m[2][0] = static_cast<float>(-sin(rad));
+	m[0][2] = static_cast<float>(sin(rad));
 
 	return m;
 }
@@ -64,9 +64,9 @@ mat4f Math::buildRotateMatrixZ(float rad)
 	mat4f m;
 	m.toIdentity();
 
-	m[0][0] = m[1][1] = cos(rad);
-	m[0][1] = -sin(rad);
-	m[1][0] = sin(rad);
+	m[0][0] = m[1][1] = static_cast<float>(cos(rad));
+	m[0][1] = static_cast<float>(-sin(rad));
+	m[1][0] = static_cast<float>(sin(rad));
 
 	return m;
 }
